+++ content/blog/happy-kodar/happy-kodar.11tydata.js
@@ -1,0 +1,6 @@
+export default {
+	permalink: function ({ page }) {
+		return `/blog/${page.fileSlug}/`;
+	},
+	layout: "layouts/happy.njk"
+};