+++ eleventy.config.js
@@ -136,6 +136,14 @@ export default async function(eleventyConfig) {
 		return (new Date()).toISOString();
 	});
 
+	eleventyConfig.addCollection("elevate", (collectionApi) => {
+		return collectionApi.getFilteredByGlob("content/blog/elevate-with-nordling/{,**/}*.md");
+	});
+
+	eleventyConfig.addCollection("happy", (collectionApi) => {
+		return collectionApi.getFilteredByGlob("content/blog/happy-kodar/{,**/}*.md");
+	});
+
 	// Features to make your build faster (when you need them)
 
 	// If your passthrough copy gets heavy and cumbersome, add this line